Compute Inertia Ratio
Inertia Ratio is recorded online in newer versions of ShapeIn. This property can also be obtained offline from the contours. Inertia Ratio is closely liked to "Deformation"  (=1-circulrity), but has the advantage to be less affected by protrusions and non-convex shapes. In future it might also be used to look up E-Modulus using an appropriate LUT.