Add "Principal Inertia Ratio" (rot. invariance)
The inertia ratio is not invariant to rotation of the contour. This means especially in the reservoir that cells which flow by in arbitrary angle get an altered inertia ratio. Interestingly, the second moment of intertia allows to compute the orientation of the object. In this current update, the user can choose 'align'=True in order to rotate the contour to align it with the coordinate system and then compute the inertia ratio based on the Principal Moment of Inertia (we could call it Principal Inertia Ratio)